Update CAN DBC definitions for motor controller integration
- Update motorControllerRequest message ID from 0x700 to 0x201
- Change requestType signal to Throttle_Value (16-bit)
- Replace motorStatus message with motorStatus_PDO1 (0x391) and motor_status_PDO4 (0x271)
- Add SDO_request message for CANopen communication
- Update SYNC frequency from 30Hz to 1Hz
- Regenerate all CAN DBC code files with new message definitions

+++ RTOS/Scheduler/SysTick.h
@@ -76,8 +76,8 @@ void SysTick_CPUTimerStart(void);
  */
 void SysTick_CPUTimerEnd(void);
 
-float SysTick_GetCPUPercentage(void);
-float SysTick_GetCPUPeak(void);
+uint16_t SysTick_GetCPUPercentage(void);
+uint16_t SysTick_GetCPUPeak(void);
 
 /**
  * @brief SysTick_GetTimerValue returns the current TMR5 value
